daemon/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pyyaml>=6.0
watchfiles>=1.0
